Add `ChatGeneration` tests and catch `DeprecationWarning`
